restore full course loader for lesson sidebar
